Session.php creation
Trying to troubleshoot welcome.php
There's a boolean in session.php which handles logging in and is supposed to say the user is logged in.

This is then passed to the header of welcome.php and i supposed to boot the users back to login.php if the boolean is false. Is automatically flipping to false regardless of what happens in session.php